moved general settings widget to specific file
